all pass and compiles

+++ MovieData.cpp
@@ -28,15 +28,15 @@ void MovieData:: UpdateMovieViewer(int i)
     this->m_views += i;
 }
 
-int MovieData:: UpdateMovieRating(int i)
+double MovieData:: UpdateMovieRating(int i)
 {
-    int add= (m_rating*m_raters+i)/(m_raters+1)- m_rating;
+    double add= (m_rating*m_raters+i)/(m_raters+1)- m_rating;
     m_rating=(m_rating*m_raters+i)/(m_raters+1);
     m_raters+=1;
     return add;
 }
 
-int MovieData::getMovieRating() const
+double MovieData::getMovieRating() const
 {
     return m_rating;
 }
